+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 "Unreleased" changes are implemented but haven't been uploaded yet.
 
-## Unreleased
+## [0.1.5] - 2017-08-29
   - Mac: Cut/Copy/Paste added to Menu. Should fix "can't copy/paste" on macOS.
   - Export to JSON feature available again.
   - Fixed "Can't add cards above/below first card".